Add test for leaking allocator during clonning
