+++ src/raw/mod.rs
@@ -3393,88 +3393,6 @@ mod test_map {
         }
     }
 
-    #[test]
-    #[should_panic = "panic in clone"]
-    fn test_clone_memory_leaks_and_double_drop() {
-        use ::alloc::vec::Vec;
-
-        const DISARMED: bool = false;
-        const ARMED: bool = true;
-
-        struct CheckedCloneDrop {
-            panic_in_clone: bool,
-            dropped: bool,
-            need_drop: Vec<i32>,
-        }
-
-        impl Clone for CheckedCloneDrop {
-            fn clone(&self) -> Self {
-                if self.panic_in_clone {
-                    panic!("panic in clone")
-                }
-                Self {
-                    panic_in_clone: self.panic_in_clone,
-                    dropped: self.dropped,
-                    need_drop: self.need_drop.clone(),
-                }
-            }
-        }
-
-        impl Drop for CheckedCloneDrop {
-            fn drop(&mut self) {
-                if self.dropped {
-                    panic!("double drop");
-                }
-                self.dropped = true;
-            }
-        }
-
-        let mut table: RawTable<(u64, CheckedCloneDrop)> = RawTable::with_capacity(7);
-
-        let armed_flags = [
-            DISARMED, DISARMED, ARMED, DISARMED, DISARMED, DISARMED, DISARMED,
-        ];
-
-        // Hash and Key must be equal to each other for controlling the elements placement
-        // so that we can be sure that we first clone elements that don't panic during cloning.
-        for (idx, &panic_in_clone) in armed_flags.iter().enumerate() {
-            let idx = idx as u64;
-            table.insert(
-                idx,
-                (
-                    idx,
-                    CheckedCloneDrop {
-                        panic_in_clone,
-                        dropped: false,
-                        need_drop: vec![0, 1, 2, 3],
-                    },
-                ),
-                |(k, _)| *k,
-            );
-        }
-        let mut count = 0;
-        unsafe {
-            // Let's check that all elements are located as we wanted
-            //
-            // SAFETY: We know for sure that `RawTable` will outlive
-            // the returned `RawIter` iterator.
-            for (bucket, panic_in_clone) in table.iter().zip(armed_flags) {
-                // SAFETY: It's OK, we get bucket from RawIter, so it is safe to read
-                let (key, value) = bucket.as_ref();
-                assert_eq!(*key, count);
-                assert_eq!(value.panic_in_clone, panic_in_clone);
-                count += 1;
-            }
-        }
-        assert_eq!(table.len(), 7);
-        assert_eq!(count, 7);
-
-        // Clone should normally clone a few elements, and then (when the clone
-        // function panics), deallocate both its own memory and the memory of
-        // already cloned elements (Vec<i32> memory inside CheckedCloneDrop).
-        let _table2 = table.clone();
-    }
-
     /// CHECKING THAT WE ARE NOT TRYING TO READ THE MEMORY OF
     /// AN UNINITIALIZED TABLE DURING THE DROP
     #[test]
